# lib/api-zod/src/generated/types/bowelMovement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BowelMovement {
  id: number;
  residentId: number;
  bristolType: number;
  amount: string;
  incontinence: boolean;
  bloodPresent: boolean;
  mucusPresent: boolean;
  painStraining: boolean;
  clinicalNote: string;
  createdAt: Date;
}