# lib/api-zod/src/generated/types/bowelMovementInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BowelMovementInputAmount } from "./bowelMovementInputAmount";

export interface BowelMovementInput {
  residentId: number;
  /**
   * @minimum 1
   * @maximum 7
   */
  bristolType: number;
  amount: BowelMovementInputAmount;
  incontinence: boolean;
  bloodPresent: boolean;
  mucusPresent: boolean;
  painStraining: boolean;
  clinicalNote: string;
}