# lib/api-zod/src/generated/types/physicianSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentAlertSummary } from "./residentAlertSummary";

export interface PhysicianSummary {
  residents: ResidentAlertSummary[];
  facilityMonthlyGaps: number;
  facilityMonthlyBlood: number;
  generatedAt: Date;
}