# lib/api-zod/src/generated/types/resident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Resident {
  id: number;
  name: string;
  room: string;
}