# lib/api-zod/src/generated/types/residentAlertSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentAlertSummaryAlertLevel } from "./residentAlertSummaryAlertLevel";

export interface ResidentAlertSummary {
  residentId: number;
  name: string;
  room: string;
  alertLevel: ResidentAlertSummaryAlertLevel;
  /** @nullable */
  lastBMAt: Date | null;
  /** @nullable */
  hoursSinceLastBM: number | null;
  monthlyGapCount: number;
  monthlyBloodCount: number;
}